caster-back/requirements.txt
django==4.1.3
uvicorn[standard]==0.17.5
gunicorn==20.1.0
psycopg2-binary==2.9.3
python-osc==1.8.0
google-cloud-texttospeech==2.10.2
strawberry-graphql-django==0.8.2
strawberry-graphql[channels]<=0.154.0
strawberry-django-plus==1.35
django-cors-headers==3.13.0
mistletoe==0.9.0
pydantic==1.10.2
